+++ randomizer/Lists/DoorLocations.py
@@ -3,7 +3,7 @@ from randomizer.Enums.Events import Events
 from randomizer.Enums.Kongs import Kongs
 from randomizer.Enums.Levels import Levels
 from randomizer.Enums.Regions import Regions
-from randomizer.Lists.MapsAndExits import Maps
+from randomizer.Enums.Maps import Maps
 from randomizer.Logic import Regions as RegionList
 from randomizer.LogicClasses import TransitionFront
