+++ randomizer/Patching/DoorPlacer.py
@@ -5,7 +5,7 @@ import js
 from randomizer.Enums.ScriptTypes import ScriptTypes
 from randomizer.Enums.Settings import MiscChangesSelected
 from randomizer.Lists.DoorLocations import door_locations
-from randomizer.Lists.MapsAndExits import Maps
+from randomizer.Enums.Maps import Maps
 from randomizer.Patching.Lib import addNewScript, float_to_hex, getNextFreeID, IsItemSelected
 from randomizer.Patching.Patcher import ROM, LocalROM
