+++ randomizer/Patching/EnemyRando.py
@@ -6,7 +6,7 @@ from randomizer.Enums.EnemySubtypes import EnemySubtype
 from randomizer.Enums.Settings import CrownEnemyRando, DamageAmount
 from randomizer.Lists.EnemyTypes import Enemies, EnemyMetaData, enemy_location_list
 from randomizer.Enums.EnemyLocations import EnemyLocations
-from randomizer.Lists.MapsAndExits import Maps
+from randomizer.Enums.Maps import Maps
 from randomizer.Patching.Patcher import ROM, LocalROM
