+++ randomizer/Patching/EntranceRando.py
@@ -2,7 +2,8 @@
 import js
 from randomizer.Enums.Settings import ShuffleLoadingZones
 from randomizer.Enums.Transitions import Transitions
-from randomizer.Lists.MapsAndExits import GetExitId, GetMapId, MapExitTable, Maps
+from randomizer.Enums.Maps import Maps
+from randomizer.Lists.MapsAndExits import GetExitId, GetMapId, MapExitTable
 from randomizer.Patching.Patcher import ROM, LocalROM
 
 valid_lz_types = [9, 12, 13, 16]
