+++ randomizer/Patching/FairyPlacer.py
@@ -4,7 +4,7 @@ import js
 from randomizer.Enums.Levels import Levels
 from randomizer.Lists.EnemyTypes import Enemies
 from randomizer.Lists.FairyLocations import fairy_locations, relocated_5ds_fairy
-from randomizer.Lists.MapsAndExits import Maps
+from randomizer.Enums.Maps import Maps
 from randomizer.Patching.Patcher import ROM, LocalROM
