+++ randomizer/Patching/KasplatLocationRando.py
@@ -2,7 +2,7 @@
 import js
 from randomizer.Lists.EnemyTypes import Enemies
 from randomizer.Lists.KasplatLocations import KasplatLocationList
-from randomizer.Lists.MapsAndExits import Maps
+from randomizer.Enums.Maps import Maps
 from randomizer.Patching.Patcher import ROM, LocalROM
