+++ randomizer/ShuffleBosses.py
@@ -8,7 +8,7 @@ from randomizer.Enums.Levels import Levels
 from randomizer.Enums.Locations import Locations
 from randomizer.Lists.Exceptions import BossOutOfLocationsException, FillException, ItemPlacementException
 from randomizer.Lists.Location import LocationList
-from randomizer.Lists.MapsAndExits import Maps
+from randomizer.Enums.Maps import Maps
 from randomizer.Enums.Settings import HardModeSelected
 from randomizer.Patching.Lib import IsItemSelected
