+++ randomizer/ShuffleShopLocations.py
@@ -5,7 +5,7 @@ import randomizer.Logic as Logic
 from randomizer.Enums.Levels import Levels
 from randomizer.Enums.Regions import Regions
 from randomizer.Enums.Settings import ShuffleLoadingZones
-from randomizer.Lists.MapsAndExits import Maps
+from randomizer.Enums.Maps import Maps
 from randomizer.LogicClasses import TransitionFront
 from randomizer.Spoiler import Spoiler
