+++ randomizer/ShuffleWarps.py
@@ -4,7 +4,7 @@ import random
 import js
 import randomizer.Logic as Logic
 from randomizer.Enums.Warps import Warps
-from randomizer.Lists.MapsAndExits import Maps
+from randomizer.Enums.Maps import Maps
 from randomizer.Lists.Warps import BananaportVanilla, VanillaBananaportSelector
 from randomizer.LogicClasses import TransitionFront
